获取国家、城市方法和数据分离
